move 'map' tests to separate file
