+++ types/ramda/tsconfig.json
@@ -124,6 +124,7 @@
         "test/lensIndex-tests.ts",
         "test/lensProp-tests.ts",
         "test/lt-tests.ts",
-        "test/lte-tests.ts"
+        "test/lte-tests.ts",
+        "test/map-tests.ts"
     ]
 }
